pyproject.toml
[tool.poetry]
name = "vran-django"
version = "0.1.0"
description = ""
authors = ["Moritz Fürneisen <m.fuerneisen@zbw.eu>"]
readme = "README.md"
packages = [{include = "vran"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
Django = "^4.1.3"
django-ninja = "^0.19.1"
django-cors-headers = "^3.13.0"
backports-zoneinfo = {markers = "python_version < \"3.9\"", version = "^0.2.1"}
pandas = "^2.0.1"
django-rq = "^2.8.1"
psycopg = "^3.1.9"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-cov = "*"
black = "*"
coverage = "*"
pdoc3 = "*"
pre-commit = "*"
isort = "*"
pylint = "*"
pytest-django = "^4.5.2"
requests = "^2.28.1"
python-dotenv = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
